update project naming
