+++ frontend/src/constants/ProjectActionTypes.js
@@ -8,3 +8,5 @@ export const TASK_CLAIMED = 'TASK_CLAIMED'
 export const TASK_COMPLETED = 'TASK_COMPLETED'
 export const TASK_VALIDATED = 'TASK_VALIDATED'
 export const UPDATE_PROJECT = 'UPDATE_PROJECT'
+export const GET_PROPOSED_PROJECTS = 'GET_PROPOSED_PROJECTS'
+export const PROPOSED_PROJECTS_RECEIVED = 'PROPOSED_PROJECTS_RECEIVED'
