--- frontend/src/constants/getters/ProjectGetterActionTypes.js
@@ -1,2 +1,0 @@
-export const GET_PROPOSED_PROJECTS = 'GET_PROPOSED_PROJECTS'
-export const PROPOSED_PROJECTS_RECEIVED = 'PROPOSED_PROJECTS_RECEIVED'
