+++ frontend/src/reducers/index.js
@@ -1,18 +1,18 @@
 import NavigationReducer from './NavigationReducer'
 import userReducer from './userReducer'
 import pollReducer from './pollReducer'
-import statusGetterReducer from './statusReducer'
-import projectGetterReducer from './projectReducer'
+import statusReducer from './statusReducer'
+import projectReducer from './projectReducer'
 import { routerReducer } from 'react-router-redux'
 import { combineReducers } from 'redux'
 
 const Nemo = combineReducers({
   navigation: NavigationReducer,
   user: userReducer,
-  projects: projectGetterReducer,
+  projects: projectReducer,
   polls: pollReducer,
   router: routerReducer,
-  status: statusGetterReducer
+  status: statusReducer
 })
 
 export default Nemo
